Goof: Refactor fundamental machinery to `pueblo_goof`
