+++ src/rapporto/cli.py
@@ -1,12 +1,12 @@
 import click
 from click_aliases import ClickAliasedGroup
-from pueblo.util.cli import boot_click
 
 from rapporto.github.cli import cli as github_cli
 from rapporto.notify.cli import cli as notify_cli
 from rapporto.opsgenie.cli import cli as opsgenie_cli
 from rapporto.report.cli import cli as report_cli
 from rapporto.slack.cli import cli as slack_cli
+from rapporto.util import setup_logging
 
 
 @click.group(cls=ClickAliasedGroup)
@@ -15,7 +15,7 @@ from rapporto.slack.cli import cli as slack_cli
 @click.version_option()
 @click.pass_context
 def cli(ctx: click.Context, verbose: bool, debug: bool):
-    return boot_click(ctx, verbose, debug)
+    setup_logging(verbose=verbose)
 
 
 cli.add_command(github_cli, "gh")
