+++ src/rapporto/notify/cli.py
@@ -2,25 +2,22 @@ import logging
 
 import click
 
-from rapporto.notify.slack.conversation import SlackConversation
-from rapporto.notify.slack.weekly import SlackWeekly
-from rapporto.util import Zapper
+from pueblo_goof.slack.conversation import SlackConversation
+from pueblo_goof.slack.model import SlackOptions, slack_api_token_option, slack_channel_option
+from pueblo_goof.util import Zapper
+from rapporto.option import github_organization_option
+from rapporto.report.model import ReportOptions
+from rapporto.report.slack import SlackWeekly
 
 logger = logging.getLogger(__name__)
 
-slack_api_token_option = click.option(
-    "--slack-token", type=str, envvar="SLACK_TOKEN", required=False, help="Slack API token"
-)
-slack_channel_option = click.option(
-    "--slack-channel", "-c", type=str, envvar="SLACK_CHANNEL", required=False, help="Slack channel"
-)
-
 
 @click.group()
+@github_organization_option
 @slack_api_token_option
 @slack_channel_option
 @click.pass_context
-def cli(ctx: click.Context, slack_token: str, slack_channel: str):
+def cli(ctx: click.Context, github_organization: str, slack_token: str, slack_channel: str):
     """
     Notify humans and machines.
     """
@@ -28,8 +25,10 @@ def cli(ctx: click.Context, slack_token: str, slack_channel: str):
         raise click.UsageError(
             "Missing option '--slack-token' or environment variable 'SLACK_TOKEN'."
         )
-    ctx.meta.update({"slack_token": slack_token})
-    ctx.meta.update({"slack_channel": slack_channel})
+    ctx.meta["slack_options"] = SlackOptions(token=slack_token, channel=slack_channel)
+    ctx.meta["report_options"] = ReportOptions(
+        github_organization=github_organization, output_format="markdown"
+    )
 
 
 @cli.command()
@@ -44,10 +43,11 @@ def weekly(ctx: click.Context, week: str, zap: str):
     zapper = Zapper(zap)
     zapper.check()
 
-    conversation = SlackConversation(
-        api_token=ctx.meta["slack_token"], channel=ctx.meta["slack_channel"]
-    )
-    section = SlackWeekly(conversation=conversation, week=week)
+    report_options: ReportOptions = ctx.meta["report_options"]
+    slack_options: SlackOptions = ctx.meta["slack_options"]
+
+    conversation = SlackConversation(options=slack_options)
+    section = SlackWeekly(week=week, options=report_options, conversation=conversation)
     section.refresh()
 
     try:
