+++ src/rapporto/option.py
@@ -1,0 +1,5 @@
+import click
+
+github_organization_option = click.option(
+    "--github-organization", "--gh-org", type=str, required=True
+)
