+++ tests/pueblo_goof/conftest.py
@@ -1,0 +1,6 @@
+import pytest
+
+
+@pytest.fixture(autouse=True)
+def reset_environment(monkeypatch):
+    monkeypatch.delenv("SLACK_TOKEN", raising=False)
